+++ sqlx-core/src/mssql/outcome.rs
@@ -1,18 +1,18 @@
 use std::iter::{Extend, IntoIterator};
 
 #[derive(Debug, Default)]
-pub struct PgDone {
+pub struct MssqlOutcome {
     pub(super) rows_affected: u64,
 }
 
-impl PgDone {
+impl MssqlOutcome {
     pub fn rows_affected(&self) -> u64 {
         self.rows_affected
     }
 }
 
-impl Extend<PgDone> for PgDone {
-    fn extend<T: IntoIterator<Item = PgDone>>(&mut self, iter: T) {
+impl Extend<MssqlOutcome> for MssqlOutcome {
+    fn extend<T: IntoIterator<Item = MssqlOutcome>>(&mut self, iter: T) {
         for elem in iter {
             self.rows_affected += elem.rows_affected;
         }
@@ -20,9 +20,9 @@ impl Extend<PgDone> for PgDone {
 }
 
 #[cfg(feature = "any")]
-impl From<PgDone> for crate::any::AnyDone {
-    fn from(done: PgDone) -> Self {
-        crate::any::AnyDone {
+impl From<MssqlOutcome> for crate::any::AnyOutcome {
+    fn from(done: MssqlOutcome) -> Self {
+        crate::any::AnyOutcome {
             rows_affected: done.rows_affected,
             last_insert_id: None,
         }