+++ sqlx-core/src/query_as.rs
@@ -84,7 +84,7 @@ where
     pub fn fetch_many<'e, 'c: 'e, E>(
         self,
         executor: E,
-    ) -> BoxStream<'e, Result<Either<DB::Done, O>, Error>>
+    ) -> BoxStream<'e, Result<Either<DB::Outcome, O>, Error>>
     where
         'q: 'e,
         E: 'e + Executor<'c, Database = DB>,