+++ sqlx-core/src/sqlite/outcome.rs
@@ -1,12 +1,12 @@
 use std::iter::{Extend, IntoIterator};
 
 #[derive(Debug, Default)]
-pub struct SqliteDone {
+pub struct SqliteOutcome {
     pub(super) changes: u64,
     pub(super) last_insert_rowid: i64,
 }
 
-impl SqliteDone {
+impl SqliteOutcome {
     pub fn rows_affected(&self) -> u64 {
         self.changes
     }
@@ -16,8 +16,8 @@ impl SqliteDone {
     }
 }
 
-impl Extend<SqliteDone> for SqliteDone {
-    fn extend<T: IntoIterator<Item = SqliteDone>>(&mut self, iter: T) {
+impl Extend<SqliteOutcome> for SqliteOutcome {
+    fn extend<T: IntoIterator<Item = SqliteOutcome>>(&mut self, iter: T) {
         for elem in iter {
             self.changes += elem.changes;
             self.last_insert_rowid = elem.last_insert_rowid;
@@ -26,9 +26,9 @@ impl Extend<SqliteDone> for SqliteDone {
 }
 
 #[cfg(feature = "any")]
-impl From<SqliteDone> for crate::any::AnyDone {
-    fn from(done: SqliteDone) -> Self {
-        crate::any::AnyDone {
+impl From<SqliteOutcome> for crate::any::AnyOutcome {
+    fn from(done: SqliteOutcome) -> Self {
+        crate::any::AnyOutcome {
             rows_affected: done.changes,
             last_insert_id: Some(done.last_insert_rowid),
         }